projekt/Backend/requirements.txt
bcrypt==5.0.0
blinker==1.9.0
click==8.1.8
colorama==0.4.6
flask==3.1.2
Flask-Bcrypt==1.0.1
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
flask-sqlalchemy==3.1.1
greenlet==3.2.4
gunicorn==23.0.0
importlib-metadata==8.7.0
itsdangerous==2.2.0
jinja2==3.1.6
markupsafe==3.0.3
packaging==25.0
PyJWT==2.10.1
sqlalchemy==2.0.44
typing-extensions==4.15.0
werkzeug==3.1.3
zipp==3.23.0
google-auth==2.43.0
requests==2.32.3
pytest==8.3.4
pytest-cov==6.0.0
